[UI] Add new Date components (Range, Single, Part)
Both Range and Single have a start date and end date in the data model.
Single date interprets the start date as the earliest moment of that
date and end date as the latest, e.g. 00.00 and 23.59.59...

For dateType "before" the start date is set as 1600-01-01.

It also clears the no longer editable dates when changing date type, to
ensure the user knows what they input.