[COMMON] Improve date range building utils
1. Rewrite buildDateRange to have no logic except building the object
and removing undefined values and in the end checking if the object has
any value, then return it else return null

2. Add getInterpretedDateRange* that looks at the dateType and decides
what to pass to the buildDateRange function. E.g. if dateType is latest
then set use getEarliestTimestamp for startInterpretedTimestamp and set
date inputs as end date. If not latest date, then set date inputs as
both start and end date inputs