[COMMON] Set relationshipsArray to [] by default
A bit dangerous commit but this is the way nestedToCore should behave.
The danger is related to a possibility of fetching a resource without
all relationships and then saving it without specifying which
relationships to edit. Then existing relationships could be
overwritten. In that case the solution is to specify which
relationships you intent to update