[ROOT] Reorder jobs in travis.yml to longest first
Travis does not always start all jobs at the same time and it seems it
goes from top to bottom. When e2e was last, that meant the longest
running job started last, which maximises the runtime. This reorders the
jobs so the approximately longest jobs come first and will be started
first.