[UI] Create different stages for tag and no tag
Travis cannot make use of cache for tag builds, because we never push
the same tag more than once and Travis identifies the cache based on the
tag. Since the lack of a cache can almost double the time a job takes to
run, it is reasonable to have specific stages optimised for running
without cache. E.g. the ui build takes almost double the time without
cache, which means we might as well do all other builds in the same job,
whereas when there is a cache it is so fast we shorten the critical path
by splitting the other builds in two jobs.

Similarly, with cache it makes sense to run all non-e2e tests in one
job, whereas without cache they should be split up to not take more time
than e2e (which is not greatly impacted by the existence of cache or
not).

Lastly, the tag builds will be faster when not unnecessarily having to
upload the (unusable) cache files, which amount to ~750 MB.