[UI] Improve number Input handling invalid input
This prevents the user from writing another minus sign or dot if the
value is a valid number with a minus or dot included. It does not
prevent all non-numeric input, as that is not possible.

Input of type
number prevents the value (the HTML property) from becoming non-numeric
by setting it to an empty string if the user input is non-numeric.
However, it does not prevent the user from typing all non-numeric
input, e.g. you can write messed up stuff like "`2" or "1---2" also
here (i.e. it is not a bug that we have introduced in our form):
https://developer.mozilla.org/en-US/docs/Web/HTML/Element/input/number

Mozilla Firefox has confirmed that browsers handle non-numeric input
differently and it is yet to be resolved:
https://bugzilla.mozilla.org/show_bug.cgi?id=1398528

React also cannot
fix it, because it's either not possible (as it relates to how browsers
implement the HTML specification) or not clear what would be a
better/less bad solution:
https://github.com/facebook/react/issues/2203#issuecomment-56969977