[UI] Add wait after storage location test
When refetching the specimen after updating storage location the
storage location is removed from the state for a short time during which
the form will show an error due to storage location being required. If
trying to navigate away during that time the prompt will prevent that,
which causes the test suite to hang. Reproduced in headless mode but not
in headful mode.